Separate 'Triplestore' Class and better API
