+++ triplestore/src/triplestore/__init__.py
@@ -1,6 +1,17 @@
 # Copyright (C) 2025 Maira Papadopoulou
 # SPDX-License-Identifier: Apache-2.0
 
-from triplestore.registration import Triplestore, available_backends
+from triplestore.base import TriplestoreBackend
+from triplestore.exceptions import BackendNotFoundError, BackendNotInstalledError, TriplestoreError, TriplestoreMissingConfigValue
+from triplestore.registration import available_backends
+from triplestore.triplestore import Triplestore
 
-__all__ = ["Triplestore", "available_backends"]
+__all__ = [
+        "BackendNotFoundError",
+        "BackendNotInstalledError",
+        "Triplestore",
+        "TriplestoreBackend",
+        "TriplestoreError",
+        "TriplestoreMissingConfigValue",
+        "available_backends",
+]
